Catch the kill process error
Print a warning message if a process kill fails.
The process kill is a safety measure against a hanging process.
At the moment all the running processes seems to run and close fine,
so we are always trying to kill an already closed process.
Some workers (WSL, msys2) running a SPSA test raise an exception that stops
the execution for some seconds slowing down the test.

Also raise version to 72.